Replace DatagramDtlsAdaptor with DatagramRtpTransport.
DatagramDtlsAdaptor wraps a DatagramTransport in a DtlsTransport.  This
is only used by wrapping it again, in an RtpTransport.  It is simpler to
just wrap DatagramTransport directly into an RtpTransport.

DatagramTransport is never used as a DtlsTransport, and doesn't support
most of the functionality exposed by the DtlsTransport interface.

However, it supports *all* the functionality of the RtpTransport, making
this a much cleaner fit.

Bug: webrtc:9719
Change-Id: I699e8124ee4cb6c8c187162f9b444ff0431a4902