pipewire: handle deleting the capturer while a D-Bus call is in progress
If a D-Bus is in progess when a BaseCapturerPipeWire is deleted, then the
user_data is invalid when the callback function is called. This results in
memory corruption.

To fix this, use a GCancellable. If it is canceled, the callback will be
called with a corresponding error. Detect this error and abort before
accessing the user_data.

pipewire: use correct argument for g_dbus_proxy_call() callback

The first argument is the 'source_object'. For g_dbus_proxy_call() this is
the proxy object not the connection.

Bug: None
Change-Id: I8d5e3fb5c49fcc9afd61cdb8e8249f78b9434faf