Reduces locking in RtpSenderVideo.
This CL replaces some locking with just a SequenceChecker.
It also removes the usage of RegisterPayloadType() and replaces it with
a paremeter in SendVideo instead. This way we are prepared for removing
the payload type map and lock entirely. Some usage still exists
downstream and needs to be removed before cleaning this up.

Bug: webrtc:10809
Change-Id: Ie90163f15d11c8843f3beaf9a0df0dd2a1fd5ce6