Reduces locking in RtpSenderVideo.
This CL some unnecessary locking, since we are already serialized by
the lock in VideoStreamEncoder. Unfortunately we can't use a normal
SequenceChecker as we don't know which thread/task queue the encoder
internal implementation decides to run the OnEncodedImage() callback on.

We also remove the usage of RegisterPayloadType() and replace it with
a parameter in SendVideo instead. This way we are prepared for removing
the payload type map and lock entirely. Some usage still exists
downstream and needs to be removed before cleaning this up.

Bug: webrtc:10809
Change-Id: Ie90163f15d11c8843f3beaf9a0df0dd2a1fd5ce6