Update patch set 3
Patch Set 3:

I agree that it makes sense to not have the static approach.

However, even though the change seems small, it totally changes the scope for all of these methods. 

Would it not instead be possible to create non-class functions for these functions so that the scope is maintained?

Patch-set: 3