Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I agree that it makes sense to not have the static approach.
> 
> However, even though the change seems small, it totally changes the scope for all of these methods. 
> 
> Would it not instead be possible to create non-class functions for these functions so that the scope is maintained?
A static member function atually has the same scope as a regular member function. The only difference is that they don't take an implicit class pointer called |this|. It's not possible to move them outside the class scope wince that would prohibit access to private members.

It's a bit confusing now, since they are static, but takes a mutable pointer to the class, making them equivalent to a member function with the only difference that we need to write stream->private_member_ instead of this->private_member_/private_member_.

Patch-set: 3