Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> Wouldn't it be simpler to just make a function that takes two async operations an returns a new async operation, like so:
> 
>   template <typename F1, typename F2>
>   auto Chain(F1&& f1, F2&& f2) {
>     return [f1 = std::forward<F1>(f1),
>             f2 = std::forward<F2>(f2)](auto&& callback) mutable {
>              f1([f2 = std::move(f2),
>                  callback = std::forward<decltype(callback)>(callback)]() mutable {
>                   f2(std::move(callback));
>                 });
>            };
>   }
> 
> Full example here: https://godbolt.org/z/J0vwZk

Given two operations, if I want to create a compound operation as two operations, I can take two functors and return chained functor. But this is a very limited use case.

What I need is a queue that I can enqueue things on at any time. When operation A starts executing the client hasn't even requested for operation B, so I can't combine A and B. Operations A and B are not aware of each other, and neither is the OperationsChain until ChainOperation() is invoked at arbitrary points in time. This is why the callback has to go back to the OperationsChain and see if there are any new operations to handle.

Patch-set: 3
Reviewer: Gerrit User 5142 <5142@58829da1-049c-39fb-b951-ebdcd0984767>