Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Don't know the details here, but if you need an owner review I'll be happy to LGTM once someone else has had a look as well. Or I get a good explanation. :)

Yes, so basically, what we are doing is that we ensure that we won't allocate audio as part of the total bitrate allocation unless we have negotiated the transport sequence number extension. Previously we incorrectly did that if the send side bwe for audio field trial was not enabled. With this change the behavior is consistent, independently of that trial, and that only regulates whether we will enable sending of the transport sequence number extension. (That's something we likely want to revisit as well, but that's for another CL)

Patch-set: 2
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>