Define WEBRTC_ENABLE_SYMBOL_EXPORT if is_component_build=true.
In order to land the component build support in Chromium, it is
easier to turn on symbols export every time that is_component_build=true
instead of setting rtc_enable_symbol_export=is_component_build in
Chromium (since is_component_build is not available in .gn).

rtc_enable_symbol_export is still kept in the mix in order to turn
on symbol exports in any case a shared library will be added to the
WebRTC build.

Bug: webrtc:9419
Change-Id: I5a7195826dea13d9a6f10a1160c35f2864bfa6c2