Account for RTX MID/RRID overhead when packetizing
The criteria for sending MID and/or RID is different for the
primary SSRC and RTX SSRC. It could be the case that MID and/or RID
is not sent on the primary SSRC but an RTX packet with MID and/or
RRID needs to be constructed for one of those packets. To allow for
that we need to leave enough room to add those header extensions in
the RTX packet without exceeding the max packet size.

Bug: webrtc:11031
Change-Id: I314a9ecca8651b32c3f57b32593da1c5e62c4088