Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> Hi Yura, thanks for the background information on the test runs. I'll let you try to uncover the seg fault before I further review. 
> 
> In the meantime I'd like to move Mirko as a reviewer to get his take on having this type of long running test in WebRTC. We do have other performance tests in WebRTC and I'm wondering if this is something that would better fit there. 
> 
> Mirko, do you have any thoughts?

First of all thanks for working on this Yura.

I also add phoglund@ for a second opinion. We have perf tests but they are already hitting the 2 hours timeout and we are going to work on them soon in order to make them run faster.

Adding a 1 hour test is tricky, I am wondering if we have to keep it as a manual test or maybe run it nightly (even if we don't have such tests at the moment).

Patrik, any opinion on this?

Patch-set: 5
Reviewer: Gerrit User 5087 <5087@58829da1-049c-39fb-b951-ebdcd0984767>