Update patch set 5
Patch Set 5:

(3 comments)

> Patch Set 5:
> 
> > We have perf tests but they are already hitting the 2 hours timeout and we are going to work on them soon in order to make them run faster.
> 
> Mirko, thank you very much for reviewing this CL! 
> I'd like to express that the test which is added in this CL is more like reliability/robustness/stress test for usrsctp, rather than a performance test in classical meaning. The most important here is to test to successfully finish, no matter an actual execution time. I think the test is closer to fuzzing tests, rather than unit or performance tests. You probably have better understanding and more info where to put this kind of tests and how often run theym. I personally used the test from CL to debug my proposed fix (https://github.com/sctplab/usrsctp/pull/397) for deadlock issue in newer version of usrsctp (https://github.com/sctplab/usrsctp/issues/325).
> Thank you very much for having a look at this CL and discussion!

Hi!

I want the tests to follow the arrange-act-assert rule, otherwise it looks nice. This is wayyyy more complex than what a unit test would allow to be, but this is an end-to-end stress test so that's fine.

We can't run a one-hour test. That's out of the question. I would say the maximum we could allow would be 60 seconds. If you want write some useful variant of the test that runs < 60s we can add it to the test suites. Do all of your new tests take one hour? We could add some of the tests if they are faster.

We have made experiments with long-running regression tests in the past but it doesn't work. They're too slow to provide actionable information if they actually break. It can work if you do something like oss-fuzz/clusterfuzz, where they run 24/7 with lots of machines and very good diagnostics (e.g. it abuses the code until it finds a memory error, which by definition is a bug and they can exit immediately with great diagnostics). Tests like these are good for targeted efforts like finding this deadlock though. We can add the long tests to the repo but they will have to be disabled until it's time to debug deadlocks again.

Patch-set: 5
Reviewer: Gerrit User 5027 <5027@58829da1-049c-39fb-b951-ebdcd0984767>