Update patch set 9
Patch Set 9:

(2 comments)

LGTM as long as there are no performance concerns related to calling FindFullScreenWindow as often as is happening right now. Please wait for an LGTM from wez@ (or other expert) regarding the Win32 code.

Also, I'm a bit concerned about the lack of tests in this code-base. The new functionality added by this CL is all platform-specific, and hence probably not unit-testable, but the refactoring does afford an opportunity to add tests for the cross-platform portions. It doesn't need to block submission of this CL, but is that something you could look into as a follow-up?

Patch-set: 9
Reviewer: Gerrit User 5166 <5166@58829da1-049c-39fb-b951-ebdcd0984767>