Update patch set 5
Patch Set 5: Commit-Queue+2

> Patch Set 5: Code-Review+1
> 
> Were no further allow_poison annotations necessary? Meaning that our production stuff does in fact not include the json library?

Yes, well, mostly correct!
The AEC3 JSON config parsing uses the library. This is an _optional_ part of WebRTC, used in Chrome to parse and apply settings for AEC3.

Projects that use the run-of-the-mill WebRTC library are free of the JSON library.
Projects that _also_ use this little JSON parsing target, are not free.

Patch-set: 5
Reviewer: Gerrit User 5119 <5119@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=+2