P2PTransportChannel - Fix needed sorting for correct Ping:ing
This patch adds a sorting of the connection list when selecting
which connection to ping. The sorting is only performed _if needed_
and is gated behind a field trial for safe rollout.

This can happen in the following scenario
a) Connection state change => RequestSortAndStateUpdate
b) The Ping selection code run and assumes a sorted list
c) The posted sort request comes in "too late".

FYI: this was found while working on https://webrtc-review.googlesource.com/c/src/+/158205

BUG=webrtc:11050

Change-Id: I908582d737b6b36bf8f9bfe58add60b35377954d