Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > > 
> > > Thank you very much for having a look at this CL again!
> > > I've already tried release builds with ASAN, MSAN and TSAN locally.
> > > With latest changes to this CL there is no issue detected by ASAN and MSAN, but TSAN detects issue inside `usrsctp` itself.
> > > The TSAN issue could only be fixed by fixing it inside `usrsctp` (https://github.com/sctplab/usrsctp), which is different repository, so I mean there is no way right now to make the test green under TSAN other than disabling test back :( But other fixes from this CL are useful, because they fix valid TSAN and MSAN error in test code itself.
> > > 
> > > Regarding the permission to try-bots. I've seen a document describing policy for Chromium project (https://www.chromium.org/getting-involved/become-a-committer), but could not find such document for WebRTC. I would be happy and it would be more convenient for me and probably for reviewers, if I could be granted a bit more permissions to WebRTC repo with my current list of contributions if there is enough of them.
> > 
> > I looked at your contributions and you have contributed a good amount of clean CLs. Congratulations, you're now a committer! That means you also have tryjob access.
> 
> Thank you very much, Patrik! I'm very happy to become a committer.
> I've checked that now I can ask try-bot for a build and it works, that's nice! 
> Does a "committer" right mean I can also send my approved CL's to CQ? Can I send this one to CQ, I mean is your LGTM is still valid after recent patches fixing MSAN and partially TSAN? 
> Could you please also share a document (if it exists) describing a committer responsibilities, which are coming with that great power? Thank you very much again, Patrik!

Yes, you can now press +1 and +2 yourself. Being a committer means you now have the power to review, edit bugs and revert patches. However, you're not in any OWNERS files AFAIK so your review isn't enough to approve a patch.

A LGTM counts for later patches as long as I don't withdraw it. If you get a LGTM and make significant changes after, it's customary to ask the reviewer to take another look (so the reviewer isn't surprised at what code is actually going in). If you do trivial changes you don't have to ask for another review.

You're expected to be a guardian of the codebase in reviews and keep it consistent and of high quality.

Patch-set: 6
Reviewer: Gerrit User 5027 <5027@58829da1-049c-39fb-b951-ebdcd0984767>