Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

> Patch Set 4: Code-Review+1
> 
> Hi Qingsi,
> 
> Sorry for late reply, I missed this yesterday 😞
> I think patch looks great, super clean.
> 
> I only have 1 comment that is about API-ish...which I'm not an expert on.
> - should the cricket::IceTransportInternal also be moved to api/ ??
> (or maybe api/ice/ so one can move more of the definintions/enum there going forward)
> 
> Go for it!

Comment on exposing cricket::IceTransportInternal:
this was a hack done to allow the present code in Blink's ice_transport_adapter_impl to continue to work.

There are 2 ways to create a webrtc::IceTransport - one is the internal one that is done by JSEP, the other (which I think is the one you call "default" now) is the one called by Blink (and only used for the experimental QuicTransport API). Unifying them, or making them into 2 versions of factories, would be great.

There are a couple of leftovers from the introduction of webrtc::IceTransport that need doing:
- Make the internal() function RTC_GUARDED_BY(network_thread). At the moment, it's available to the signaling thread because the JSEP transport controller, which runs on the signaling thread, takes a cricket::IceTransportInternal pointer and posts it to the networking thread - now that this patch makes the controller use webrtc::IceTransportInterface instead, this should be revisited.
- Remove the internal() function from Blink, and instead expose the functions that the Blink code needs on the webrtc::IceTransportInterface API (or move them to the init function that Blink uses).


As far as I can tell, this refactoring / extension makes that simpler, not harder, so I like it!

Patch-set: 4
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1