Update patch set 2
Patch Set 2: Commit-Queue+1

This an attempt to reland part of the
https://webrtc-review.googlesource.com/c/src/+/159034
that was the reason of the rollback.

The difference is that this time priority of the TQ is higher. May that would avoid the issue.

Patch-set: 2
Reviewer: Gerrit User 5117 <5117@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5019 <5019@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=+1
Work-in-progress: false