Update patch set 3
Patch Set 3:

I'll mostly quibble about the name.

This is a time series moving average. The most common moving average assumes constant time between samples (or samples that are a value series without a relationship with time).

Either the documentation or the class name (or preferably both) should reflect this.

I am not statistician enough to understand how the time-series effect will affect the statistics of the result, or whether it matters for the applications we're thinking of. But it is important to call out.

(I think Niels has done a good job of commenting on the code already.)

Patch-set: 3
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>