Update patch set 6
Patch Set 6:

(2 comments)

Thanks for the quick review! Adding kwiberg to reviewer list.

Hoping to land this step ASAP (peerconnection is so big that it changes frequently, and this CL touches a lot of it, so rebasing is a PITA)

Patch-set: 6
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>