Update patch set 2
Patch Set 2:

> Expected result after step 6 will be rotated AppRTCMobile app in landscape mode.

If you've hung up the call, we should stop listening to orientation changes right? We are only interested during the call. But I think you had an idea that the begin- method would be called multiple times if the user switches camera, so the calls end up unbalanced? I just wanted to make sure there is actually a repro case where the calls are unbalanced and it leads to a problem in practice. Thanks!

Patch-set: 2
Reviewer: Gerrit User 5249 <5249@58829da1-049c-39fb-b951-ebdcd0984767>