Update patch set 1
Patch Set 1:

Since this is a API/behavioral change, I think we should have an explicit bug for it to make it easier to trace down and investigate in case it causes downstream issues.

Essentially, a bug describing the current design, why this is bad, and how it should work instead.

Patch-set: 1
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>