Make VideoStreamEncoder responsible for reconfiguring its source/sink.
Prior to this CL the VideoSourceProxy (residing inside
OveruseFrameDetectorResourceAdaptationModule) was responsible for
performing AddOrUpdateSink(). This could happen for various reasons:
- AdaptUp() or AdaptDown() modified the "sink wants", informing the
  VideoStreamEncoder of changes to VideoSourceRestrictions, which then
  asked the module to reconfigure the VideoSourceProxy.
- The VideoStreamEncoder asked to reconfigure the source for reasons
  other than restrictions changing, such as
  SetMaxFramerateAndAlignment() and SetWantsRotationApplied(). These
  requests were plumbed through the adaptation module even though they
  were unrelated to resource adaptation.

(WIP, TODO):

After this CL, VideoSourceController has replaced VideoSourceProxy. The
controller is owned by the VideoStreamEncoder which is now directly
responsible for any reconfiguration of the video source. This means
that the OveruseFrameDetectorResourceAdaptationModule is no longer
aware of any settings other than VideoSourceRestrictions.

Bug: webrtc:11222
Change-Id: I0f058c4700ca108e2d9f212e38b61f6f728aa419