Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > nisse@ told me that the compiler would syntehsize the move operators if I =default'ed the copy operators.
> > 
> > https://webrtc-review.googlesource.com/c/src/+/161304/7/api/rtc_error.h#122
> > 
> > Nisse, please confirm.
> 
> According to https://stackoverflow.com/questions/12073556/c-11-is-a-defaulted-copy-constructor-user-declared which quotes the standard marking a copy constructor as =default makes it 'user-declared' which prevents all implicit constructors from being defined.
> 
> So in order to get default copy & move operations, we should either mark them all explicitly as =default (this CL) or remove all the =default declarations.

OK, then this CL has LGTM from me.

The C++ style guide says that we should be explicit about whether or not the type is copyable and/or moveable - that is, either declare them or delete them: https://google.github.io/styleguide/cppguide.html#Copyable_Movable_Types
WebRTC fails to do that for a lot of classes, I guess we should clean that up as a drive-by when we change them for other reasons - now I know what is needed in that case!

Patch-set: 1
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1