Update patch set 1
Patch Set 1:

On second reading, I agree with Mirko: A compile time constant is not the Right Thing.

If this is really needed, we should expose a setter in the API that allows the same flexibility as the Hayes spec (1-255 seconds), so that clients can modify this at runtime, not compile time.

Patch-set: 1
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>