Update patch set 2
Patch Set 2:

I've modified the patch set to allow users to specify a comma gap delay other than the spec compliant value of 2 seconds. I opted to enable this via an additional parameter to InsertDtmf with a default value of 2 seconds so that existing clients would not have to update and to be consistent with how the inter_tone_gap and duration are currently set.

Patch-set: 2
Reviewer: Gerrit User 7217 <7217@58829da1-049c-39fb-b951-ebdcd0984767>