Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

Looks good, just a few nits.

This is nice in that it will have no effect on people not using it, and as far as I can tell, the value will be reset to default whenever InsertDtmf is called without the extra argument.

Please fix the nits, but doing LGTM to allow you to submit when ready.

Patch-set: 2
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1